+++ src/main/java/com/ll/project_13_backend/payment/service/CashService.java
@@ -4,9 +4,7 @@ import com.ll.project_13_backend.global.BaseEntity;
 import com.ll.project_13_backend.member.entity.Member;
 import com.ll.project_13_backend.payment.entity.CashLog;
 
-import java.math.BigDecimal;
-
 public interface CashService {
 
-    CashLog addCash(Member member, BigDecimal price, CashLog.EvenType eventType, BaseEntity relEntity) ;
+    CashLog addCash(Member member, long price, CashLog.EvenType eventType, BaseEntity relEntity) ;
 }
