+++ src/main/java/com/ll/project_13_backend/payment/service/CashServiceImpl.java
@@ -8,8 +8,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.math.BigDecimal;
-
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
@@ -18,7 +16,7 @@ public class CashServiceImpl implements CashService{
     private final CashLogRepository cashLogRepository;
 
     @Transactional
-    public CashLog addCash(Member member, BigDecimal price, CashLog.EvenType eventType, BaseEntity relEntity) {
+    public CashLog addCash(Member member, long price, CashLog.EvenType eventType, BaseEntity relEntity) {
         CashLog cashLog = CashLog.builder()
                 .member(member)
                 .price(price)
