+++ src/main/java/com/ll/project_13_backend/post/service/PostService.java
@@ -3,12 +3,10 @@ package com.ll.project_13_backend.post.service;
 import com.ll.project_13_backend.member.entity.Member;
 import com.ll.project_13_backend.post.entity.Category;
 import com.ll.project_13_backend.post.entity.Post;
-
-import java.math.BigDecimal;
 public interface PostService {
 
 
-    Post createPost(Member member, String title, String content, Category category, BigDecimal price);
+    Post createPost(Member member, String title, String content, Category category, long price);
 
 
 }
