+++ src/main/java/com/ll/project_13_backend/post/service/PostServiceImpl.java
@@ -8,14 +8,12 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.math.BigDecimal;
-
 @Service
 @RequiredArgsConstructor
 public class PostServiceImpl implements PostService {
     private final PostRepository postRepository;
     @Transactional
-    public Post createPost(Member member, String title, String content, Category category, BigDecimal price) {
+    public Post createPost(Member member, String title, String content, Category category, long price) {
 
       Post post = Post.builder()
               .member(member)
